Backend_new/requirements.txt
fastapi
python-multipart
uvicorn
PyAudio
pygame
groq
python-dotenv
aiofiles
markdown
google-genai
openai